+++ tests/testthat/test-aggregates.R
@@ -5,9 +5,6 @@ library(matsbyname)
 library(matsindf)
 library(tidyr)
 
-###########################################################
-context("Aggregates")
-###########################################################
 
 test_that("primary aggregates of SUT data work as expected", {
   # Define primary industries
@@ -16,8 +13,10 @@ test_that("primary aggregates of SUT data work as expected", {
   # Primary TOTAL aggregates
   primary_total_aggregates_sut <- UKEnergy2000mats %>%
     tidyr::spread(key = matrix.name, value = matrix) %>%
-    primary_aggregates(p_industries = p_industries, by = "Total",
-                       aggregate_primary = "EX_total_agg.ktoe")
+    dplyr::mutate(
+      p_industries = rep(list(p_industries), times = nrow(.))
+    ) %>%
+    primary_aggregates(p_industries = "p_industries", by = "Total", aggregate_primary = "EX_total_agg.ktoe")
   expect_equivalent(primary_total_aggregates_sut %>% dplyr::filter(Last.stage == IEATools::last_stages$final) %>%
                       dplyr::select("EX_total_agg.ktoe"), 93000)
   expect_equivalent(primary_total_aggregates_sut %>% dplyr::filter(Last.stage == IEATools::last_stages$useful) %>%
@@ -30,7 +29,10 @@ test_that("primary aggregates of SUT data work as expected", {
   # Primary PRODUCT aggregates
   primary_product_aggregates_sut <- UKEnergy2000mats %>%
     tidyr::spread(key = matrix.name, value = matrix) %>%
-    primary_aggregates(p_industries = p_industries, by = "Product",
+    dplyr::mutate(
+      p_industries = rep(list(p_industries), times = nrow(.))
+    ) %>%
+    primary_aggregates(p_industries = "p_industries", by = "Product",
                        aggregate_primary = "EX_product_agg.ktoe") %>%
     dplyr::select(Country, Year, Last.stage, Energy.type, EX_product_agg.ktoe) %>%
     tidyr::gather(key = "matnames", value = "matvals", EX_product_agg.ktoe) %>%
@@ -47,7 +49,10 @@ test_that("primary aggregates of SUT data work as expected", {
   # Primary FLOW aggregates
   primary_flow_aggregates_sut <- UKEnergy2000mats %>%
     tidyr::spread(key = matrix.name, value = matrix) %>%
-    primary_aggregates(p_industries = p_industries, by = "Flow",
+    dplyr::mutate(
+      p_industries = rep(list(p_industries), times = nrow(.))
+    ) %>%
+    primary_aggregates(p_industries = "p_industries", by = "Flow",
                        aggregate_primary = "EX_flow_agg.ktoe") %>%
     dplyr::select(Country, Year, Last.stage, Energy.type, EX_flow_agg.ktoe) %>%
     tidyr::gather(key = "matnames", value = "matvals", EX_flow_agg.ktoe) %>%
@@ -68,15 +73,76 @@ test_that("primary aggregates work when R is folded into V", {
 
   # Primary TOTAL aggregates
   primary_total_aggregates_sut <- UKEnergy2000mats %>%
+    # Make it wide-by-matrices
     tidyr::spread(key = matrix.name, value = matrix) %>%
     dplyr::mutate(
+      p_industries = rep(list(p_industries), times = nrow(.)),
       V = matsbyname::sum_byname(R, V),
       R = NULL
     ) %>%
-    primary_aggregates(p_industries = p_industries, by = "Total",
+    primary_aggregates(p_industries = "p_industries", by = "Total",
                        aggregate_primary = "EX_total_agg.ktoe")
   expect_equivalent(primary_total_aggregates_sut %>% dplyr::filter(Last.stage == IEATools::last_stages$final) %>%
                       dplyr::select("EX_total_agg.ktoe"), 93000)
+})
+
+
+test_that("primary aggregates work when p_industries is different for each row of a data frame", {
+  p_industries_all <- c("Resources - Crude", "Resources - NG")
+  p_industries_crude <- "Resources - Crude"
+  p_industries_ng <- "Resources - NG"
+
+  UKEnergy2000mats_wide <- UKEnergy2000mats %>%
+    # Make it wide-by-matrices
+    tidyr::pivot_wider(names_from = "matrix.name", values_from = "matrix")
+  # Primary TOTAL aggregates for both primary industries
+  primary_total_aggregates_both <- UKEnergy2000mats_wide %>%
+    dplyr::mutate(
+      p_industries = rep(list(p_industries_all), times = nrow(UKEnergy2000mats_wide))
+    ) %>%
+    primary_aggregates(p_industries = "p_industries", by = "Total",
+                       aggregate_primary = Recca::aggregate_cols$aggregate_primary)
+  expect_equivalent(primary_total_aggregates_both %>% dplyr::filter(Last.stage == IEATools::last_stages$final) %>%
+                      dplyr::select(Recca::aggregate_cols$aggregate_primary), 93000)
+
+
+  # Primary TOTAL aggregates counting crude oil only
+  primary_total_aggregates_crude <-UKEnergy2000mats_wide %>%
+    dplyr::mutate(
+      p_industries = p_industries_crude
+    ) %>%
+    primary_aggregates(p_industries = "p_industries", by = "Total",
+                       aggregate_primary = Recca::aggregate_cols$aggregate_primary)
+  expect_equivalent(primary_total_aggregates_crude %>% dplyr::filter(Last.stage == IEATools::last_stages$final) %>%
+                      dplyr::select(Recca::aggregate_cols$aggregate_primary), 50000)
+
+  # Primary TOTAL aggregates counting NG only
+  primary_total_aggregates_ng <- UKEnergy2000mats_wide %>%
+    dplyr::mutate(
+      p_industries = p_industries_ng
+    ) %>%
+    primary_aggregates(p_industries = "p_industries", by = "Total",
+                       aggregate_primary = Recca::aggregate_cols$aggregate_primary)
+  expect_equivalent(primary_total_aggregates_ng %>% dplyr::filter(Last.stage == IEATools::last_stages$final) %>%
+                      dplyr::select(Recca::aggregate_cols$aggregate_primary), 43000)
+
+  # Try when the vectors of primary industries are different down the column in the data frame
+  UK_final_only <- UKEnergy2000mats_wide %>%
+    dplyr::filter(.data[[IEATools::iea_cols$last_stage]] == IEATools::last_stages$final)
+  UK_final_3_rows <- dplyr::bind_rows(UK_final_only, UK_final_only, UK_final_only) %>%
+  # Add a column containing various options for primary industries to be aggregated.
+  dplyr::mutate(
+    p_industries = list(p_industries_all, p_industries_crude, p_industries_ng)
+  )
+  res <- UK_final_3_rows %>%
+    primary_aggregates(p_industries = "p_industries", by = "Total",
+                       aggregate_primary = Recca::aggregate_cols$aggregate_primary)
+  # Check first row (should have both primary industries and have 93000 for the aggregate primary energy)
+  expect_equal(res[[Recca::aggregate_cols$aggregate_primary]][[1]], 93000)
+  # Check second row (should have only crude and have 50000 for the aggregate primary energy)
+  expect_equal(res[[Recca::aggregate_cols$aggregate_primary]][[2]], 50000)
+  # Check third row (should have only NG and have 43000 for the aggregate primary energy)
+  expect_equal(res[[Recca::aggregate_cols$aggregate_primary]][[3]], 43000)
 })
 
 
@@ -87,10 +153,13 @@ test_that("final demand aggregates of SUT data work as expected", {
   # Final demand TOTAL aggregates
   final_demand_total_aggregates_sut <- UKEnergy2000mats %>%
     tidyr::spread(key = matrix.name, value = matrix) %>%
-    finaldemand_aggregates_with_units(fd_sectors = fd_sectors, by = "Total",
+    dplyr::mutate(
+      fd_sectors = rep(list(fd_sectors), times = nrow(.))
+    ) %>%
+    finaldemand_aggregates_with_units(fd_sectors = "fd_sectors", by = "Total",
                        net_aggregate_demand = "EX_total_net_agg.ktoe",
                        gross_aggregate_demand = "EX_total_gross_agg.ktoe") %>%
-    dplyr::select(Country, Year, Last.stage, Energy.type, EX_total_net_agg.ktoe, EX_total_gross_agg.ktoe) %>%
+    dplyr::select(Country, Year, Last.stage, Energy.type, fd_sectors, EX_total_net_agg.ktoe, EX_total_gross_agg.ktoe) %>%
     tidyr::gather(key = "matnames", value = "matvals", EX_total_net_agg.ktoe, EX_total_gross_agg.ktoe) %>%
     matsindf::expand_to_tidy(drop = 0)
   expect_equivalent(final_demand_total_aggregates_sut %>%
@@ -113,7 +182,10 @@ test_that("final demand aggregates of SUT data work as expected", {
   # Final demand PRODUCT aggregates
   final_demand_product_aggregates_sut <- UKEnergy2000mats %>%
     tidyr::spread(key = matrix.name, value = matrix) %>%
-    finaldemand_aggregates_with_units(fd_sectors = fd_sectors, by = "Product",
+    dplyr::mutate(
+      fd_sectors = rep(list(fd_sectors), times = nrow(.))
+    ) %>%
+    finaldemand_aggregates_with_units(fd_sectors = "fd_sectors", by = "Product",
                                       net_aggregate_demand = "EX_product_net_agg.ktoe",
                                       gross_aggregate_demand = "EX_product_gross_agg.ktoe") %>%
     dplyr::select(Country, Year, Last.stage, Energy.type, EX_product_net_agg.ktoe, EX_product_gross_agg.ktoe) %>%
@@ -147,7 +219,10 @@ test_that("final demand aggregates of SUT data work as expected", {
   # Final demand SECTOR aggregates
   final_demand_sector_aggregates_sut <- UKEnergy2000mats %>%
     tidyr::spread(key = matrix.name, value = matrix) %>%
-    finaldemand_aggregates_with_units(fd_sectors = fd_sectors, by = "Sector",
+    dplyr::mutate(
+      fd_sectors = rep(list(fd_sectors), times = nrow(.))
+    ) %>%
+    finaldemand_aggregates_with_units(fd_sectors = "fd_sectors", by = "Sector",
                                       net_aggregate_demand = "EX_sector_net_agg.ktoe",
                                       gross_aggregate_demand = "EX_sector_gross_agg.ktoe") %>%
     dplyr::select(Country, Year, Last.stage, Energy.type, EX_sector_net_agg.ktoe, EX_sector_gross_agg.ktoe) %>%
@@ -179,54 +254,40 @@ test_that("final demand aggregates of SUT data work as expected", {
                     142916629629)
 })
 
-test_that("primary_aggregates_IEA works as expected", {
-  # Get a vector of primary industries for the example data set.
-  # The vector of primary industries comes from the resource matrix (R).
-  r_ind <- UKEnergy2000mats %>%
-    tidyr::spread(key = matrix.name, value = matrix) %>%
-    extract2("R") %>%
-    extract2(1) %>%
-    rownames()
 
+test_that("IEATools::primary_aggregates_IEA() works as expected on the UKEnergy2000tidy data frame", {
   result <- UKEnergy2000tidy %>%
-    dplyr::group_by(Country, Year, Energy.type, Last.stage) %>%
-    primary_aggregates_IEA(p_industries = r_ind)
-  expect_equal(result[["EX_p_IEA.ktoe"]], c(93000, 93000, 93000, 98220))
+    dplyr::mutate(
+      Method = "PCM"
+    ) %>%
+    dplyr::group_by(Country, Method, Year, Energy.type, Last.stage) %>%
+    IEATools::primary_aggregates()
+  expect_equal(result[[Recca::aggregate_cols$aggregate_primary]], c(93000, 93000, 93000, 98220))
 })
 
-test_that("finaldemand_aggregates_IEA works as expected", {
-  iea_result <- UKEnergy2000tidy %>%
-    # Can calculate only when all entries are in same units, i.e., only when last stage is final or useful energy.
-    dplyr::filter(Last.stage %in% c(IEATools::last_stages$final, IEATools::last_stages$useful)) %>%
-    dplyr::group_by(Country, Year, Energy.type, Last.stage) %>%
-    finaldemand_aggregates_IEA()
-  sut_result <- UKEnergy2000mats %>%
-    tidyr::spread(key = matrix.name, value = matrix) %>%
-    dplyr::filter(Last.stage %in% c(IEATools::last_stages$final, IEATools::last_stages$useful)) %>%
-    finaldemand_aggregates(fd_sectors = c("Residential", "Transport"))
-  expect_equal(iea_result[["EX_fd_net_IEA.ktoe"]], sut_result[["EX_fd_net.ktoe"]] %>% unlist())
-  expect_equal(iea_result[["EX_fd_gross_IEA.ktoe"]], sut_result[["EX_fd_gross.ktoe"]] %>% unlist())
-})
 
 test_that("finaldemand_aggregates works for sectors", {
   sut_result <- UKEnergy2000mats %>%
     tidyr::spread(key = matrix.name, value = matrix) %>%
+    dplyr::mutate(
+      fd_sectors = rep(list(c("Residential", "Transport")), times = nrow(.))
+    ) %>%
     dplyr::filter(Last.stage %in% c(IEATools::last_stages$final, IEATools::last_stages$useful)) %>%
-    finaldemand_aggregates(fd_sectors = c("Residential", "Transport"), by = "Sector")
-  expect_equal(sut_result$EX_fd_net.ktoe[[1]][1,1], 31000)
-  expect_equal(sut_result$EX_fd_net.ktoe[[1]][2,1], 40750)
-  expect_equal(sut_result$EX_fd_net.ktoe[[2]][1,1], 4200.4)
-  expect_equal(sut_result$EX_fd_net.ktoe[[2]][2,1], 21714.9805)
+    finaldemand_aggregates(fd_sectors = "fd_sectors", by = "Sector")
+  expect_equal(sut_result[[Recca::aggregate_cols$net_aggregate_demand]][[1]][1,1], 31000)
+  expect_equal(sut_result[[Recca::aggregate_cols$net_aggregate_demand]][[1]][2,1], 40750)
+  expect_equal(sut_result[[Recca::aggregate_cols$net_aggregate_demand]][[2]][1,1], 4200.4)
+  expect_equal(sut_result[[Recca::aggregate_cols$net_aggregate_demand]][[2]][2,1], 21714.9805)
 
-  expect_equal(sut_result$EX_fd_gross.ktoe[[1]][1,1], 550)
-  expect_equal(sut_result$EX_fd_gross.ktoe[[1]][2,1], 350)
-  expect_equal(sut_result$EX_fd_gross.ktoe[[1]][4,1], 2075)
-  expect_equal(sut_result$EX_fd_gross.ktoe[[1]][11,1], 40750)
+  expect_equal(sut_result[[Recca::aggregate_cols$gross_aggregate_demand]][[1]][1,1], 550)
+  expect_equal(sut_result[[Recca::aggregate_cols$gross_aggregate_demand]][[1]][2,1], 350)
+  expect_equal(sut_result[[Recca::aggregate_cols$gross_aggregate_demand]][[1]][4,1], 2075)
+  expect_equal(sut_result[[Recca::aggregate_cols$gross_aggregate_demand]][[1]][11,1], 40750)
 
-  expect_equal(sut_result$EX_fd_gross.ktoe[[2]][1,1], 0)
-  expect_equal(sut_result$EX_fd_gross.ktoe[[2]][2,1], 45)
-  expect_equal(sut_result$EX_fd_gross.ktoe[[2]][6,1], 75)
-  expect_equal(sut_result$EX_fd_gross.ktoe[[2]][11,1], 26.9997)
-  expect_equal(sut_result$EX_fd_gross.ktoe[[2]][14,1], 21714.9805)
+  expect_equal(sut_result[[Recca::aggregate_cols$gross_aggregate_demand]][[2]][1,1], 0)
+  expect_equal(sut_result[[Recca::aggregate_cols$gross_aggregate_demand]][[2]][2,1], 45)
+  expect_equal(sut_result[[Recca::aggregate_cols$gross_aggregate_demand]][[2]][6,1], 75)
+  expect_equal(sut_result[[Recca::aggregate_cols$gross_aggregate_demand]][[2]][11,1], 26.9997)
+  expect_equal(sut_result[[Recca::aggregate_cols$gross_aggregate_demand]][[2]][14,1], 21714.9805)
 })